requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
lancedb
wikipedia
arxiv
langchainhub
groq
langchain-groq
beautifulsoup4
